+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `destroy` function to an instanced model.
 
+### Changed
+- Turn off writing to depth buffer when rendering transparent objects using `StandardMaterial`. This feature if only available in PixiJS 6.0+.
+
 ### Fixed
 - Fixed an issue which caused mesh instances to be rendered even if they were `visible = false`.
 - Fixed an issue which caused `StandardMaterial` not to render double sided materials correctly.